requirements.txt
strands-agents
strands-agents-tools
google-genai
python-dotenv
pyyaml
rich
temporalio
